requirements.txt
# Existing dependencies (from your current system)
av>=14.4.0
pymediainfo>=6.0.0
PyExifTool>=0.5.0
opencv-python>=4.8.0
typer[all]>=0.9.0
rich>=13.4.0
pydantic>=2.4.0
structlog>=23.1.0
numpy>=1.24.0
pillow>=10.0.0
polyfile>=0.5.5
hachoir==3.3.0
python-dateutil>=2.8.2
transformers>=4.28.0
torch>=2.0.0
google-genai
python-dotenv

# Flask dependencies for API server
flask
flask-cors>=4.0.0

# New Supabase dependencies
tiktoken>=0.5.0
openai

# Prefect dependencies
prefect>=2.0.0

# Local Database & Testing Dependencies
duckdb>=1.0.0
pytest>=7.0.0   # For running tests